requirements.txt
opencv-python
opencv-contrib-python
numpy
pickle-mixin
Pillow
tk
pyinstaller